fix: BF16 AMP에서 cross_entropy dtype 불일치 수정
tag_logits.float()로 캐스팅하여 ce_weight(float32)와 dtype 일치.
mamba_ssm fused kernel → split path 전환 후 autocast가 BF16 출력을 반환하면서 발생.

+++ training/pretrain_dense_editor.py
@@ -158,7 +158,7 @@ def validate_editor(model, val_loader, criterion, config, device, use_amp, n_ste
             targets = edit_tags.clone()
             targets[~pad_mask] = -100
             loss = criterion(
-                tag_logits.view(-1, config.n_tags),
+                tag_logits.float().view(-1, config.n_tags),
                 targets.view(-1),
             )
 
@@ -541,7 +541,7 @@ def train(args):
                 targets = torch.where(pad_mask, edit_tags, _ignore_idx)
 
                 ce_loss = criterion(
-                    tag_logits.view(-1, config.n_tags),
+                    tag_logits.float().view(-1, config.n_tags),
                     targets.view(-1),
                 )
